heavy optimizations important change: do not wake up if only a volume button is pressed under the hood: - flag bits in GPIOR0 instead of variables - low_byte in GPIOR1 - smaller type of some global variables

+++ wavfile.h
@@ -18,9 +18,9 @@ typedef struct {
 } WAVfmtChunkData_t;
 
 typedef struct {
-	uint16_t	num_channels;
-	uint32_t	sample_rate;
-	uint16_t	bits_per_sample;
+	uint8_t		num_channels;
+	uint16_t	sample_rate;
+	uint8_t		bits_per_sample;
 	uint8_t		block_align;
 	uint32_t	data_length;
 } WAVinfo_t;
